MB1-141 Refactor: 초기 유저 설정 기본값 mutation을 app라우터로 이동
