+++ src/pages/Calendar/index.tsx
@@ -1,13 +1,9 @@
-import { useEffect } from "react";
 import { CalendarList, DateData, LocaleConfig } from "react-native-calendars";
 import MaterialIcons from "@expo/vector-icons/MaterialIcons";
-import { useIsNewUserContext } from "@/src/contexts/IsNewUserProvider";
-import useRegisterSetting from "@/src/hooks/useRegisterSetting";
-import Loading from "@/src/components/ui/Loading";
-import CommonError from "@/src/components/ui/CommonError";
 import DayComponent from "./DayComponent";
 import { theme } from "@/src/constants/theme";
 import * as S from "./styles";
+import { getCalendarFn } from "@/src/apis/calendarApi";
 
 LocaleConfig.locales["ko"] = {
   monthNames: [
@@ -32,9 +28,6 @@ LocaleConfig.locales["ko"] = {
 LocaleConfig.defaultLocale = "ko";
 
 function CalendarPage(): JSX.Element {
-  const isNewUser = useIsNewUserContext();
-  const { isPending, isError, mutate } = useRegisterSetting();
-
   const handleCopyPress = () => {
     console.log("copy button");
   };
@@ -46,28 +39,6 @@ function CalendarPage(): JSX.Element {
   const handleDayPress = (day: DateData) => {
     console.log("selected day", day);
   };
-
-  useEffect(() => {
-    if (isNewUser) {
-      mutate();
-    }
-  }, [isNewUser, mutate]);
-
-  if (isNewUser) {
-    if (isPending) {
-      return <Loading />;
-    }
-
-    if (isError) {
-      return (
-        <CommonError
-          titleText="앱 초기 로딩 중 에러가 발생했어요."
-          buttonText="다시 로딩하기"
-          onPress={() => mutate()}
-        />
-      );
-    }
-  }
 
   return (
     <S.SafeView>
